backend/requirements.txt
django==5.1.3
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
